pyproject.toml
[project]
name = "swsabox"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytesseract (>=0.3.13,<0.4.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pygetwindow (>=0.0.9,<0.0.10)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "mss (>=10.1.0,<11.0.0)"
]

[tool.poetry]
packages = [{include = "swsabox", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
